+++ src/client.rs
@@ -2511,7 +2511,8 @@ async fn emit_event_or_shutdown(
     clippy::panic,
     clippy::todo,
     clippy::unimplemented,
-    clippy::indexing_slicing
+    clippy::indexing_slicing,
+    clippy::arithmetic_side_effects
 )]
 mod tests {
     use super::*;